v0.15.17-dev | stairs fixed && other corrections
